fix build, remove nag message to stderr when in -nc
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@6536 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/loggers/mod_logfile/mod_logfile.c
@@ -214,7 +214,7 @@ static switch_status_t process_node(const switch_log_node_t *node, switch_log_le
     
     for (hi = switch_hash_first(NULL, profile_hash); hi; hi = switch_hash_next(hi)) {
         size_t mask = 0;
-        int ok = 0;
+        size_t ok = 0;
     
         switch_hash_this(hi, &var, NULL, &val);
         profile = val;
